+++ dashboard-app/src/auth/pages/ForgotPassword.tsx
@@ -4,7 +4,8 @@ import * as Joi from 'joi';
 import { withRouter, RouteComponentProps } from 'react-router-dom';
 import { Grid, Row, Col } from 'react-flexbox-grid';
 import { Formik, Form as _Form, FormikActions } from 'formik';
-import { CbAdmins, Response } from '../../api';
+import { CbAdmins } from '../../api';
+import { Response } from '../../util/response';
 import Input from '../../components/Input';
 import { SubmitButton } from '../../components/Buttons';
 import { redirectOnError, withParams } from '../../util/routing';
