+++ dashboard-app/src/auth/pages/Login/index.tsx
@@ -4,7 +4,8 @@ import { Link as L, withRouter, RouteComponentProps } from 'react-router-dom';
 import { Grid, Row, Col } from 'react-flexbox-grid';
 import { Form as _Form, FormikActions } from 'formik';
 import { Notification } from 'react-notification';
-import { CbAdmins, Response } from '../../../api';
+import { CbAdmins } from '../../../api';
+import { Response } from '../../../util/response';
 import LoginForm, { FormValues } from './LoginForm';
 import { H1, H4 as _H4 } from '../../../components/Headings';
 import { redirectOnError, getQueryObjectFromProps } from '../../../util/routing';
