+++ dashboard-app/src/auth/pages/ResetPassword.tsx
@@ -4,7 +4,8 @@ import * as Joi from 'joi';
 import { withRouter, RouteComponentProps } from 'react-router-dom';
 import { Grid, Row, Col } from 'react-flexbox-grid';
 import { Formik, Form as _Form } from 'formik';
-import { CbAdmins, Response } from '../../api';
+import { CbAdmins } from '../../api';
+import { Response } from '../../util/response';
 import Input from '../../components/Input';
 import { SubmitButton } from '../../components/Buttons';
 import { redirectOnError, withParams, getQueryObjectFromProps } from '../../util/routing';
